--- app/controllers/api/v1/rooms_controller.rb
@@ -1,12 +1,0 @@
-module Api
-  module V1
-    class RoomsController < ApplicationController
-      def index
-        @current_user = current_user
-        redirect_to '/signin' unless @current_user
-        @rooms = Room.all
-        @users = User.all
-      end
-    end
-  end
-end
