+++ app/serializers/room_serializer.rb
@@ -1,0 +1,4 @@
+class RoomSerializer
+  include JSONAPI::Serializer
+  attributes :name
+end
