+++ config/routes.rb
@@ -7,7 +7,7 @@ Rails.application.routes.draw do
   get '/signin', to: 'sessions#new'
   post '/signin', to: 'sessions#create'
   delete '/signout', to: 'sessions#destroy'
-
+  resources :rooms
   # get "/", to: "application#welcome"
   namespace :api do
     namespace :v1 do
@@ -21,7 +21,7 @@ Rails.application.routes.draw do
         resources :event_users, only: %i[index]
       end
       resources :friendships, only: %i[create]
-      resources :rooms
+      
     end
   end
 end
