+++ spec/factories/rooms.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
   factory :room do
-    name { "MyString" }
+    name { Faker::Hipster.word }
   end
 end
