+++ spec/fixtures/vcr_cassettes/PATCH_Update_User_request/requests_to_update_an_existing_user.yml
@@ -21,7 +21,7 @@ http_interactions:
       Server:
       - nginx
       Date:
-      - Sat, 29 Jul 2023 20:23:03 GMT
+      - Sun, 30 Jul 2023 23:43:03 GMT
       Content-Type:
       - application/json; charset=UTF-8
       Transfer-Encoding:
@@ -38,7 +38,7 @@ http_interactions:
       encoding: ASCII-8BIT
       string: !binary |-
         W3sicGxhY2VfaWQiOjI4NTk5MzI5NSwibGljZW5jZSI6IkRhdGEgwqkgT3BlblN0cmVldE1hcCBjb250cmlidXRvcnMsIE9EYkwgMS4wLiBodHRwczovL29zbS5vcmcvY29weXJpZ2h0Iiwib3NtX3R5cGUiOiJ3YXkiLCJvc21faWQiOjk4MjM2MTAzNCwiYm91bmRpbmdib3giOlsiMzEuODYyMzQzNCIsIjMxLjg2MjgzOTIiLCIxMC45ODY3MDg2IiwiMTAuOTg3MDE2MyJdLCJsYXQiOiIzMS44NjI2MzYxNSIsImxvbiI6IjEwLjk4NjgxNTA2MjU3NTYzNSIsImRpc3BsYXlfbmFtZSI6IjEyMywg2YbYp9mE2YjYqiwg2LTYp9ix2Lkg2KPZgdix2YrZgtmK2KcsIE5hbHV0LCBMaWJ5YSIsImNsYXNzIjoicGxhY2UiLCJ0eXBlIjoic3F1YXJlIiwiaW1wb3J0YW5jZSI6MC4yMzUwMDAwMDAwMDAwMDAwMSwiYWRkcmVzcyI6eyJzcXVhcmUiOiIxMjMiLCJyZXNpZGVudGlhbCI6ItmG2KfZhNmI2KoiLCJzdWJ1cmIiOiLYtNin2LHYuSDYo9mB2LHZitmC2YrYpyIsImNpdHkiOiJOYWx1dCIsInN0YXRlIjoiTmFsdXQiLCJJU08zMTY2LTItbHZsNCI6IkxZLU5MIiwiY291bnRyeSI6IkxpYnlhIiwiY291bnRyeV9jb2RlIjoibHkifX1d
-  recorded_at: Sat, 29 Jul 2023 20:23:03 GMT
+  recorded_at: Sun, 30 Jul 2023 23:43:03 GMT
 - request:
     method: get
     uri: https://nominatim.openstreetmap.org/search?accept-language=en&addressdetails=1&format=json&q=123%20street,%2087654
@@ -60,7 +60,7 @@ http_interactions:
       Server:
       - nginx
       Date:
-      - Sat, 29 Jul 2023 20:23:03 GMT
+      - Sun, 30 Jul 2023 23:43:03 GMT
       Content-Type:
       - application/json; charset=UTF-8
       Transfer-Encoding:
@@ -77,5 +77,5 @@ http_interactions:
       encoding: ASCII-8BIT
       string: !binary |-
         W3sicGxhY2VfaWQiOjI4NTk5MzI5NSwibGljZW5jZSI6IkRhdGEgwqkgT3BlblN0cmVldE1hcCBjb250cmlidXRvcnMsIE9EYkwgMS4wLiBodHRwczovL29zbS5vcmcvY29weXJpZ2h0Iiwib3NtX3R5cGUiOiJ3YXkiLCJvc21faWQiOjk4MjM2MTAzNCwiYm91bmRpbmdib3giOlsiMzEuODYyMzQzNCIsIjMxLjg2MjgzOTIiLCIxMC45ODY3MDg2IiwiMTAuOTg3MDE2MyJdLCJsYXQiOiIzMS44NjI2MzYxNSIsImxvbiI6IjEwLjk4NjgxNTA2MjU3NTYzNSIsImRpc3BsYXlfbmFtZSI6IjEyMywg2YbYp9mE2YjYqiwg2LTYp9ix2Lkg2KPZgdix2YrZgtmK2KcsIE5hbHV0LCBMaWJ5YSIsImNsYXNzIjoicGxhY2UiLCJ0eXBlIjoic3F1YXJlIiwiaW1wb3J0YW5jZSI6MC4yMzUwMDAwMDAwMDAwMDAwMSwiYWRkcmVzcyI6eyJzcXVhcmUiOiIxMjMiLCJyZXNpZGVudGlhbCI6ItmG2KfZhNmI2KoiLCJzdWJ1cmIiOiLYtNin2LHYuSDYo9mB2LHZitmC2YrYpyIsImNpdHkiOiJOYWx1dCIsInN0YXRlIjoiTmFsdXQiLCJJU08zMTY2LTItbHZsNCI6IkxZLU5MIiwiY291bnRyeSI6IkxpYnlhIiwiY291bnRyeV9jb2RlIjoibHkifX1d
-  recorded_at: Sat, 29 Jul 2023 20:23:03 GMT
+  recorded_at: Sun, 30 Jul 2023 23:43:03 GMT
 recorded_with: VCR 6.2.0
