+++ spec/fixtures/vcr_cassettes/POST_Create_User_request/requests_to_create_a_new_user.yml
@@ -2,7 +2,7 @@
 http_interactions:
 - request:
     method: get
-    uri: https://nominatim.openstreetmap.org/search?accept-language=en&addressdetails=1&format=json&q=900%20Teddy%20Ways,%2094595
+    uri: https://nominatim.openstreetmap.org/search?accept-language=en&addressdetails=1&format=json&q=5897%20Orlando%20Shores,%2013495-3441
     body:
       encoding: US-ASCII
       string: ''
@@ -21,45 +21,7 @@ http_interactions:
       Server:
       - nginx
       Date:
-      - Sat, 29 Jul 2023 19:50:44 GMT
-      Content-Type:
-      - application/json; charset=UTF-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Keep-Alive:
-      - timeout=20
-      Access-Control-Allow-Origin:
-      - "*"
-      Access-Control-Allow-Methods:
-      - OPTIONS,GET
-    body:
-      encoding: UTF-8
-      string: "[]"
-  recorded_at: Sat, 29 Jul 2023 19:50:44 GMT
-- request:
-    method: get
-    uri: https://nominatim.openstreetmap.org/search?accept-language=en&addressdetails=1&format=json&q=51454%20Granville%20Island,%2027642-6070
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Server:
-      - nginx
-      Date:
-      - Sat, 29 Jul 2023 19:58:50 GMT
+      - Sun, 30 Jul 2023 23:43:01 GMT
       Content-Type:
       - application/json; charset=UTF-8
       Transfer-Encoding:
@@ -75,6 +37,44 @@ http_interactions:
     body:
       encoding: ASCII-8BIT
       string: !binary |-
-        W3sicGxhY2VfaWQiOjI5OTU3NzQ3MCwibGljZW5jZSI6IkRhdGEgwqkgT3BlblN0cmVldE1hcCBjb250cmlidXRvcnMsIE9EYkwgMS4wLiBodHRwczovL29zbS5vcmcvY29weXJpZ2h0Iiwib3NtX3R5cGUiOiJyZWxhdGlvbiIsIm9zbV9pZCI6MTIwODkxNjcsImJvdW5kaW5nYm94IjpbIjQ2LjQxMjMxMzQiLCI0Ni40NDQxMDk1IiwiLTYzLjQ5NzYyNDUiLCItNjMuNDUwNjE5MSJdLCJsYXQiOiI0Ni40MjU4NTAyIiwibG9uIjoiLTYzLjQ2OTg1NTYiLCJkaXNwbGF5X25hbWUiOiJHcmFudmlsbGUsIFF1ZWVucyBDb3VudHksIFByaW5jZSBFZHdhcmQgSXNsYW5kLCBDYW5hZGEiLCJjbGFzcyI6ImJvdW5kYXJ5IiwidHlwZSI6ImFkbWluaXN0cmF0aXZlIiwiaW1wb3J0YW5jZSI6MC41NiwiaWNvbiI6Imh0dHBzOi8vbm9taW5hdGltLm9wZW5zdHJlZXRtYXAub3JnL3VpL21hcGljb25zL3BvaV9ib3VuZGFyeV9hZG1pbmlzdHJhdGl2ZS5wLjIwLnBuZyIsImFkZHJlc3MiOnsidmlsbGFnZSI6IkdyYW52aWxsZSIsImNvdW50eSI6IlF1ZWVucyBDb3VudHkiLCJzdGF0ZSI6IlByaW5jZSBFZHdhcmQgSXNsYW5kIiwiSVNPMzE2Ni0yLWx2bDQiOiJDQS1QRSIsImNvdW50cnkiOiJDYW5hZGEiLCJjb3VudHJ5X2NvZGUiOiJjYSJ9fV0=
-  recorded_at: Sat, 29 Jul 2023 19:58:50 GMT
+        W3sicGxhY2VfaWQiOjExMDQ2NDc0MCwibGljZW5jZSI6IkRhdGEgwqkgT3BlblN0cmVldE1hcCBjb250cmlidXRvcnMsIE9EYkwgMS4wLiBodHRwczovL29zbS5vcmcvY29weXJpZ2h0Iiwib3NtX3R5cGUiOiJ3YXkiLCJvc21faWQiOjI0MTE5MDg2LCJib3VuZGluZ2JveCI6WyI0MC42NTgxNTM2IiwiNDAuNjU4MzIzIiwiLTczLjQ0NTY3MSIsIi03My40NDMwODcxIl0sImxhdCI6IjQwLjY1ODMyMyIsImxvbiI6Ii03My40NDU2NzEiLCJkaXNwbGF5X25hbWUiOiJXZXN0IE9ybGFuZG8gU3RyZWV0LCBOYXNzYXUgU2hvcmVzLCBFYXN0IE1hc3NhcGVxdWEsIFRvd24gb2YgT3lzdGVyIEJheSwgTmFzc2F1IENvdW50eSwgTmV3IFlvcmssIDExNzU4LCBVbml0ZWQgU3RhdGVzIiwiY2xhc3MiOiJoaWdod2F5IiwidHlwZSI6InJlc2lkZW50aWFsIiwiaW1wb3J0YW5jZSI6MC4zLCJhZGRyZXNzIjp7InJvYWQiOiJXZXN0IE9ybGFuZG8gU3RyZWV0IiwibmVpZ2hib3VyaG9vZCI6Ik5hc3NhdSBTaG9yZXMiLCJ0b3duIjoiRWFzdCBNYXNzYXBlcXVhIiwibXVuaWNpcGFsaXR5IjoiVG93biBvZiBPeXN0ZXIgQmF5IiwiY291bnR5IjoiTmFzc2F1IENvdW50eSIsInN0YXRlIjoiTmV3IFlvcmsiLCJJU08zMTY2LTItbHZsNCI6IlVTLU5ZIiwicG9zdGNvZGUiOiIxMTc1OCIsImNvdW50cnkiOiJVbml0ZWQgU3RhdGVzIiwiY291bnRyeV9jb2RlIjoidXMifX0seyJwbGFjZV9pZCI6MTEwNjMwMDc0LCJsaWNlbmNlIjoiRGF0YSDCqSBPcGVuU3RyZWV0TWFwIGNvbnRyaWJ1dG9ycywgT0RiTCAxLjAuIGh0dHBzOi8vb3NtLm9yZy9jb3B5cmlnaHQiLCJvc21fdHlwZSI6IndheSIsIm9zbV9pZCI6MjQxMjkxMTAsImJvdW5kaW5nYm94IjpbIjQwLjY1ODMyMyIsIjQwLjY1ODM4MiIsIi03My40NDY1NDUiLCItNzMuNDQ1NjcxIl0sImxhdCI6IjQwLjY1ODMyMyIsImxvbiI6Ii03My40NDU2NzEiLCJkaXNwbGF5X25hbWUiOiJPcmxhbmRvIFN0cmVldCwgTmFzc2F1IFNob3JlcywgRWFzdCBNYXNzYXBlcXVhLCBUb3duIG9mIE95c3RlciBCYXksIE5hc3NhdSBDb3VudHksIE5ldyBZb3JrLCAxMTc1OCwgVW5pdGVkIFN0YXRlcyIsImNsYXNzIjoiaGlnaHdheSIsInR5cGUiOiJyZXNpZGVudGlhbCIsImltcG9ydGFuY2UiOjAuMywiYWRkcmVzcyI6eyJyb2FkIjoiT3JsYW5kbyBTdHJlZXQiLCJuZWlnaGJvdXJob29kIjoiTmFzc2F1IFNob3JlcyIsInRvd24iOiJFYXN0IE1hc3NhcGVxdWEiLCJtdW5pY2lwYWxpdHkiOiJUb3duIG9mIE95c3RlciBCYXkiLCJjb3VudHkiOiJOYXNzYXUgQ291bnR5Iiwic3RhdGUiOiJOZXcgWW9yayIsIklTTzMxNjYtMi1sdmw0IjoiVVMtTlkiLCJwb3N0Y29kZSI6IjExNzU4IiwiY291bnRyeSI6IlVuaXRlZCBTdGF0ZXMiLCJjb3VudHJ5X2NvZGUiOiJ1cyJ9fSx7InBsYWNlX2lkIjoxNTQ5MTY2ODUsImxpY2VuY2UiOiJEYXRhIMKpIE9wZW5TdHJlZXRNYXAgY29udHJpYnV0b3JzLCBPRGJMIDEuMC4gaHR0cHM6Ly9vc20ub3JnL2NvcHlyaWdodCIsIm9zbV90eXBlIjoid2F5Iiwib3NtX2lkIjoyMDc4MDc4NTQsImJvdW5kaW5nYm94IjpbIjI4LjYwMDQ3MzUiLCIyOC42MDQ5OTU3IiwiLTgxLjM4NTQ4OTMiLCItODEuMzg1MTQyNyJdLCJsYXQiOiIyOC42MDI3OTQ2IiwibG9uIjoiLTgxLjM4NTMwOTciLCJkaXNwbGF5X25hbWUiOiJXeW1vcmUgUm9hZCwgV2ludGVyIFBhcmssIE9yYW5nZSBDb3VudHksIEZsb3JpZGEsIDMyODA0LCBVbml0ZWQgU3RhdGVzIiwiY2xhc3MiOiJoaWdod2F5IiwidHlwZSI6InNlY29uZGFyeSIsImltcG9ydGFuY2UiOjAuMDk5OTk5OTk5OTk5OTk5OTgsImFkZHJlc3MiOnsicm9hZCI6Ild5bW9yZSBSb2FkIiwic3VidXJiIjoiV2ludGVyIFBhcmsiLCJjb3VudHkiOiJPcmFuZ2UgQ291bnR5Iiwic3RhdGUiOiJGbG9yaWRhIiwiSVNPMzE2Ni0yLWx2bDQiOiJVUy1GTCIsInBvc3Rjb2RlIjoiMzI4MDQiLCJjb3VudHJ5IjoiVW5pdGVkIFN0YXRlcyIsImNvdW50cnlfY29kZSI6InVzIn19LHsicGxhY2VfaWQiOjE1NTIxODMwMSwibGljZW5jZSI6IkRhdGEgwqkgT3BlblN0cmVldE1hcCBjb250cmlidXRvcnMsIE9EYkwgMS4wLiBodHRwczovL29zbS5vcmcvY29weXJpZ2h0Iiwib3NtX3R5cGUiOiJ3YXkiLCJvc21faWQiOjIwNzgwNzg1MSwiYm91bmRpbmdib3giOlsiMjguNTk4OTI4NSIsIjI4LjYwMDQ3MzUiLCItODEuMzg1NTgwOCIsIi04MS4zODU0ODkzIl0sImxhdCI6IjI4LjU5OTk5NjYiLCJsb24iOiItODEuMzg1NTIwMSIsImRpc3BsYXlfbmFtZSI6Ild5bW9yZSBSb2FkLCBPcmFuZ2UgQ291bnR5LCBGbG9yaWRhLCAzMjgwNCwgVW5pdGVkIFN0YXRlcyIsImNsYXNzIjoiaGlnaHdheSIsInR5cGUiOiJzZWNvbmRhcnkiLCJpbXBvcnRhbmNlIjowLjA5OTk5OTk5OTk5OTk5OTk4LCJhZGRyZXNzIjp7InJvYWQiOiJXeW1vcmUgUm9hZCIsImNvdW50eSI6Ik9yYW5nZSBDb3VudHkiLCJzdGF0ZSI6IkZsb3JpZGEiLCJJU08zMTY2LTItbHZsNCI6IlVTLUZMIiwicG9zdGNvZGUiOiIzMjgwNCIsImNvdW50cnkiOiJVbml0ZWQgU3RhdGVzIiwiY291bnRyeV9jb2RlIjoidXMifX1d
+  recorded_at: Sun, 30 Jul 2023 23:43:01 GMT
+- request:
+    method: get
+    uri: https://nominatim.openstreetmap.org/search?accept-language=en&addressdetails=1&format=json&q=843%20O%27Conner%20Stream,%2044884
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Server:
+      - nginx
+      Date:
+      - Sun, 30 Jul 2023 23:47:42 GMT
+      Content-Type:
+      - application/json; charset=UTF-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Keep-Alive:
+      - timeout=20
+      Access-Control-Allow-Origin:
+      - "*"
+      Access-Control-Allow-Methods:
+      - OPTIONS,GET
+    body:
+      encoding: UTF-8
+      string: "[]"
+  recorded_at: Sun, 30 Jul 2023 23:47:42 GMT
 recorded_with: VCR 6.2.0
