+++ spec/requests/rooms/index_spec.rb
@@ -1,0 +1,14 @@
+# require "rails_helper"
+
+# describe "Rooms index request", :vcr do
+#   let!(:user_1) { create(:user) }
+#   let!(:room_1) { create(:room)}
+#   let!(:room_2) { create(:room)}
+#   let!(:room_3) { create(:room)}
+
+#   it "requests all the rooms" do
+#     get rooms_path
+
+#     expect(page).to have_content(room_1.name)
+#   end
+# end
